Extract entry types into new namespace for further re-usage
