--- src/core/etl/src/Flow/ETL/Row/Entry/TypedCollection/Type.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Flow\ETL\Row\Entry\TypedCollection;
-
-interface Type
-{
-    public function isEqual(self $type) : bool;
-
-    /**
-     * @param array<mixed> $collection
-     */
-    public function isValid(array $collection) : bool;
-
-    public function toString() : string;
-}
