+++ src/core/etl/tests/Flow/ETL/Tests/Unit/Row/Schema/Constraint/NotEmptyTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Flow\ETL\Tests\Unit\Row\Schema\Constraint;
 
 use Flow\ETL\DSL\Entry;
+use Flow\ETL\PHP\Type\ScalarType;
 use Flow\ETL\Row\Entries;
-use Flow\ETL\Row\Entry\TypedCollection\ScalarType;
 use Flow\ETL\Row\Schema\Constraint\NotEmpty;
 use PHPUnit\Framework\TestCase;
